+++ src/main/java/step3/controller/LottoController.java
@@ -6,6 +6,7 @@ import step3.model.LottoNumber;
 import step3.model.dto.LottoResultDto;
 import step3.model.dto.LottosNumberDto;
 import step3.view.LottoConsoleView;
+import step3.view.LottoOutputView;
 
 import java.util.List;
 
@@ -13,16 +14,16 @@ public class LottoController {
 
 
     public void start() {
-        int amount = LottoConsoleView.printPurchasingAmount();
+        int amount = LottoConsoleView.inputPurchasingAmount();
 
         LottoMachine lottoMachine = new LottoMachine(amount);
         LottosNumberDto lottosNumberDto = lottoMachine.getLottoNumber();
-        LottoConsoleView.printPurchasingLottos(lottosNumberDto);
+        LottoOutputView.printPurchasingLottos(lottosNumberDto);
 
-        String winningNumber = LottoConsoleView.printWinningNumber();
+        String winningNumber = LottoConsoleView.inputWinningNumber();
         List<LottoNumber> lottoNumbers = LottoInputParser.parseToLottoNumberArray(winningNumber);
         LottoResultDto lottoResultDto = lottoMachine.getLottoResult(lottoNumbers);
-        LottoConsoleView.printWinStats(lottoResultDto);
+        LottoOutputView.printWinStats(lottoResultDto);
     }
